requirements.txt
backports.ssl-match-hostname==3.5.0.1
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
futures==3.2.0
google-auth==1.4.1
idna==2.6
ipaddress==1.0.19
kubernetes===6.0.0
neo4j-driver==1.5.3
oauthlib==2.0.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pykka==1.2.1
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
urllib3==1.22
websocket-client==0.47.0
wheel==0.24.0
